[fix] #12 GIFImageView 분리해 중복 코드 줄이기
