+++ SniffMeet/SniffMeet/Source/Common/Helper/Extension + GIFImageView.swift
@@ -1,0 +1,39 @@
+//
+//  Extension + GIFImageView.swift
+//  SniffMeet
+//
+//  Created by 배현진 on 2/13/25.
+//
+
+import UIKit
+
+extension UIImageView {
+    func createGIFImageView(named gifName: String) -> UIImageView? {
+        guard let gifPath = Bundle.main.path(forResource: gifName, ofType: "gif"),
+              let gifData = try? Data(contentsOf: URL(fileURLWithPath: gifPath)),
+              let source = CGImageSourceCreateWithData(gifData as CFData, nil) else {
+            return nil
+        }
+
+        var images = [UIImage]()
+        var duration: Double = 0
+
+        let frameCount = CGImageSourceGetCount(source)
+        for frame in 0..<frameCount {
+            guard let cgImage = CGImageSourceCreateImageAtIndex(source, frame, nil) else { continue }
+            images.append(UIImage(cgImage: cgImage))
+
+            let frameProperties = CGImageSourceCopyPropertiesAtIndex(source, frame, nil) as? [String: Any]
+            let gifProperties = frameProperties?[kCGImagePropertyGIFDictionary as String] as? [String: Any]
+            let frameDuration = gifProperties?[kCGImagePropertyGIFDelayTime as String] as? Double ?? 0
+            duration += frameDuration
+        }
+
+        let imageView = UIImageView()
+        imageView.animationImages = images
+        imageView.animationDuration = duration
+        imageView.startAnimating()
+
+        return imageView
+    }
+}
