+++ SniffMeet/SniffMeet/Source/OnBoarding/MainBoarding/View/OnBoardingPageViewController.swift
@@ -52,7 +52,7 @@ class OnBoardingPageViewController: BaseViewController {
         )
         descriptionLabel.text = page.description
         if page.isGif {
-            if let gifImageView = createGIFImageView(named: page.imageName) {
+            if let gifImageView = UIImageView().createGIFImageView(named: page.imageName) {
                 imageView.removeFromSuperview()
                 imageView = gifImageView
                 imageView.contentMode = .scaleAspectFit
@@ -119,35 +119,6 @@ private extension OnBoardingPageViewController {
 }
 
 private extension OnBoardingPageViewController {
-    func createGIFImageView(named gifName: String) -> UIImageView? {
-        guard let gifPath = Bundle.main.path(forResource: gifName, ofType: "gif"),
-              let gifData = try? Data(contentsOf: URL(fileURLWithPath: gifPath)),
-              let source = CGImageSourceCreateWithData(gifData as CFData, nil) else {
-            return nil
-        }
-
-        var images = [UIImage]()
-        var duration: Double = 0
-
-        let frameCount = CGImageSourceGetCount(source)
-        for frame in 0..<frameCount {
-            guard let cgImage = CGImageSourceCreateImageAtIndex(source, frame, nil) else { continue }
-            images.append(UIImage(cgImage: cgImage))
-
-            let frameProperties = CGImageSourceCopyPropertiesAtIndex(source, frame, nil) as? [String: Any]
-            let gifProperties = frameProperties?[kCGImagePropertyGIFDictionary as String] as? [String: Any]
-            let frameDuration = gifProperties?[kCGImagePropertyGIFDelayTime as String] as? Double ?? 0
-            duration += frameDuration
-        }
-
-        let imageView = UIImageView()
-        imageView.animationImages = images
-        imageView.animationDuration = duration
-        imageView.startAnimating()
-
-        return imageView
-    }
-
     private func getAttributedText(
         fullText: String,
         highlight: String,
